EVG-00000 break up rules file
