+++ evergreen_lint/rules/required_expansions_write.py
@@ -1,226 +1,8 @@
-"""Lint rules."""
-import re
-from typing import Dict, List, Optional, Set, Type
+from typing import List, Optional, Set
 
-import evergreen_lint.helpers as helpers
-from evergreen_lint.helpers import (
-    determine_dependencies_of_task_def,
-    iterate_command_lists,
-    iterate_commands,
-    iterate_fn_calls_context,
-)
-from evergreen_lint.model import LintError, Rule
-
-
-class LimitKeyvalInc(Rule):
-    """Prevent/Limit usage of keyval.inc."""
-
-    @staticmethod
-    def name() -> str:
-        return "limit-keyval-inc"
-
-    @staticmethod
-    def defaults() -> dict:
-        return {"limit": 0}
-
-    def __call__(self, config: dict, yaml: dict) -> List[LintError]:
-        def _out_message(context: str) -> LintError:
-            return (
-                f"{context} uses keyval.inc. The entire file must not use "
-                f"keyval.inc more than {config['limit']} times."
-            )
-
-        out: List[LintError] = []
-        count = 0
-        for context, command in iterate_commands(yaml):
-            if "command" in command and command["command"] == "keyval.inc":
-                out.append(_out_message(context))
-                count += 1
-
-        if count <= config["limit"]:
-            return []
-        return out
-
-
-class ShellExecExplicitShell(Rule):
-    """Require explicitly specifying shell in uses of shell.exec."""
-
-    @staticmethod
-    def name() -> str:
-        return "shell-exec-explicit-shell"
-
-    @staticmethod
-    def defaults() -> dict:
-        return {}
-
-    def __call__(self, config: dict, yaml: dict) -> List[LintError]:
-        def _out_message(context: str) -> LintError:
-            return (
-                f"{context} is a shell.exec command without an explicitly "
-                "declared shell. You almost certainly want to add 'shell: "
-                "bash' to the parameters list."
-            )
-
-        out: List[LintError] = []
-        for context, command in iterate_commands(yaml):
-            if "command" in command and command["command"] == "shell.exec":
-                if "params" not in command or "shell" not in command["params"]:
-                    out.append(_out_message(context))
-
-        return out
-
-
-SHELL_COMMANDS = ["subprocess.exec", "subprocess.scripting", "shell.exec"]
-
-
-class NoWorkingDirOnShell(Rule):
-    """Do not allow working_dir to be set on shell.exec, subprocess.*."""
-
-    @staticmethod
-    def name() -> str:
-        return "no-working-dir-on-shell"
-
-    @staticmethod
-    def defaults() -> dict:
-        return {}
-
-    def __call__(self, config: dict, yaml: dict) -> List[LintError]:
-        def _out_message(context: str, cmd: str) -> LintError:
-            return (
-                f"{context} is a {cmd} command with a working_dir "
-                "parameter. Do not set working_dir, instead `cd` into the "
-                "directory in the shell script."
-            )
-
-        out: List[LintError] = []
-        for context, command in iterate_commands(yaml):
-            if "command" in command and command["command"] in SHELL_COMMANDS:
-                if "params" in command and "working_dir" in command["params"]:
-                    out.append(_out_message(context, command["command"]))
-
-        return out
-
-
-class InvalidFunctionName(Rule):
-    """Enforce naming convention on functions."""
-
-    @staticmethod
-    def name() -> str:
-        return "invalid-function-name"
-
-    @staticmethod
-    def defaults() -> dict:
-        return {"regex": "^f_[a-z][A-Za-z0-9_]*"}
-
-    def __call__(self, config: dict, yaml: dict) -> List[LintError]:
-        FUNCTION_NAME = config["regex"]
-        FUNCTION_NAME_RE = re.compile(FUNCTION_NAME)
-
-        def _out_message(context: str) -> LintError:
-            return f"Function '{context}' must have a name matching '{FUNCTION_NAME}'"
-
-        if "functions" not in yaml:
-            return []
-
-        out: List[LintError] = []
-        for fname in yaml["functions"].keys():
-            if not FUNCTION_NAME_RE.fullmatch(fname):
-                out.append(_out_message(fname))
-
-        return out
-
-
-class NoShellExec(Rule):
-    """Do not allow shell.exec. Users should use subprocess.exec instead."""
-
-    @staticmethod
-    def name() -> str:
-        return "no-shell-exec"
-
-    @staticmethod
-    def defaults() -> dict:
-        return {}
-
-    def __call__(self, config: dict, yaml: dict) -> List[LintError]:
-        def _out_message(context: str) -> LintError:
-            return (
-                f"{context} is a shell.exec command, which is forbidden. "
-                "Extract your shell script out of the YAML and into a .sh file "
-                "in directory 'evergreen', and use subprocess.exec instead."
-            )
-
-        out: List[LintError] = []
-        for context, command in iterate_commands(yaml):
-            if "command" in command and command["command"] == "shell.exec":
-                out.append(_out_message(context))
-        return out
-
-
-class NoMultilineExpansionsUpdate(Rule):
-    @staticmethod
-    def name() -> str:
-        return "no-multiline-expansions-update"
-
-    @staticmethod
-    def defaults() -> dict:
-        return {}
-
-    def __call__(self, config: dict, yaml: dict) -> List[LintError]:
-        """Forbid multi-line values in expansion.updates parameters."""
-
-        def _out_message(context: str, idx: int) -> LintError:
-            return (
-                f"{context}, key-value pair {idx} is an expansions.update "
-                "command with multi-line values embedded in the yaml, which is"
-                " forbidden. For long-form values, use the files parameter of "
-                "expansions.update."
-            )
-
-        out: List[LintError] = []
-        for context, command in iterate_commands(yaml):
-            if "command" in command and command["command"] == "expansions.update":
-                if "params" in command and "updates" in command["params"]:
-                    for idx, item in enumerate(command["params"]["updates"]):
-                        if "value" in item and "\n" in item["value"]:
-                            out.append(_out_message(context, idx))
-        return out
-
-
-class InvalidBuildParameter(Rule):
-    """Require that parameters obey a naming convention and have a description."""
-
-    @staticmethod
-    def name() -> str:
-        return "invalid-build-parameter"
-
-    @staticmethod
-    def defaults() -> dict:
-        return {"regex": "[a-z][a-z0-9_]*", "require-description": True}
-
-    def __call__(self, config: dict, yaml: dict) -> List[LintError]:
-        BUILD_PARAMETER = config["regex"]
-        BUILD_PARAMETER_RE = re.compile(BUILD_PARAMETER)
-
-        def _out_message_key(idx: int) -> LintError:
-            return f"Build parameter, pair {idx}, key must match '{BUILD_PARAMETER}'."
-
-        def _out_message_description(idx: int) -> LintError:
-            return f"Build parameter, pair {idx}, must have a description."
-
-        if "parameters" not in yaml:
-            return []
-
-        out: List[LintError] = []
-        for idx, param in enumerate(yaml["parameters"]):
-            if "key" not in param or not BUILD_PARAMETER_RE.fullmatch(param["key"]):
-                out.append(_out_message_key(idx))
-            if (
-                config["require-description"]
-                and "description" not in param
-                or not param["description"]
-            ):
-                out.append(_out_message_description(idx))
-        return out
+from evergreen_lint import helpers as helpers
+from evergreen_lint.helpers import iterate_command_lists, iterate_fn_calls_context
+from evergreen_lint.model import Rule, LintError
 
 
 class RequiredExpansionsWrite(Rule):
@@ -450,93 +232,3 @@ class RequiredExpansionsWrite(Rule):
                 out.append(_out_message_dangerous_function(context))
 
         return out
-
-
-class DependencyForFunc(Rule):
-    """
-    Define dependencies that are required if a function is used.
-
-    The configuration will look like:
-
-    ```
-    - rule: "dependency-for-func"
-      dependencies:
-        func_name: [dependency_0, depdendency_1]
-    ```
-
-    """
-
-    @staticmethod
-    def name() -> str:
-        return "dependency-for-func"
-
-    @staticmethod
-    def defaults() -> dict:
-        return {"dependencies": {}}
-
-    def __call__(self, config: dict, yaml: dict) -> List[LintError]:
-        error_msg = (
-            "Missing dependency. The task '{task_name}' expects '{dependency}' to be "
-            "listed as a dependency due to the use of the '{function}' func."
-        )
-        failed_checks = []
-        dependency_map = config.get("dependencies", {})
-        for task_def in yaml.get("tasks"):
-            actual_dependencies = determine_dependencies_of_task_def(task_def)
-            funcs = [cmd["func"] for cmd in task_def.get("commands", []) if "func" in cmd]
-            for func in funcs:
-                expected_dependencies = dependency_map.get(func, [])
-                unmet_dependenices = [
-                    dep for dep in expected_dependencies if dep not in actual_dependencies
-                ]
-                failed_checks.extend(
-                    [
-                        error_msg.format(task_name=task_def["name"], dependency=dep, function=func)
-                        for dep in unmet_dependenices
-                    ]
-                )
-
-        return failed_checks
-
-
-RULES: Dict[str, Type[Rule]] = {
-    "limit-keyval-inc": LimitKeyvalInc,
-    "shell-exec-explicit-shell": ShellExecExplicitShell,
-    "no-working-dir-on-shell": NoWorkingDirOnShell,
-    "invalid-function-name": InvalidFunctionName,
-    "no-shell-exec": NoShellExec,
-    "no-multiline-expansions-update": NoMultilineExpansionsUpdate,
-    "invalid-build-parameter": InvalidBuildParameter,
-    "required-expansions-write": RequiredExpansionsWrite,
-    "dependency-for-func": DependencyForFunc,
-}
-# Thoughts on Writing Rules
-# - see .helpers for reliable iteration helpers
-# - Do not assume a key exists, unless it's been mentioned here
-# - Do not allow exceptions to percolate outside of the rule function
-# - YAML anchors are not available. Unless you want to write your own yaml
-#   parser, or fork adrienverge/yamllint, abandon all hope on that idea you have.
-# - Anchors are basically copy and paste, so you might see "duplicate" errors
-#   that originate from the same anchor, but are reported in multiple locations
-
-# Evergreen YAML Root Structure Reference
-# Unless otherwise mentioned, the key is optional. You can infer the
-# substructure by reading etc/evergreen.yml
-
-# Function blocks: are dicts with the key 'func', which maps to a string,
-# the name of the function
-# Command blocks: are dicts with the key 'command', which maps to a string,
-# the Evergreen command to run
-
-# variables: List[dict]. These can be any valid yaml and it's very difficult
-#   to infer anything
-# functions: Dict[str, Union[dict, List[dict]]]. The key is the name of the
-#   function, the value is either a dict, or list of dicts, with each dict
-#   representing a command
-# pre, post, and timeout: List[dict] representing commands or functions to
-#   be run before/after/on timeout condition respectively
-# tasks: List[dict], each dict is a task definition, key is always present
-# task_groups: List[dict]
-# modules: List[dict]
-# buildvariants: List[dict], key is always present
-# parameters: List[dict]
